requirement.txt
asgiref==3.5.2
Django==4.1.3
django-debug-toolbar==3.7.0
django-environ==0.9.0
djangorestframework==3.14.0
mysqlclient==2.1.1
psycopg2==2.9.5
psycopg2-binary==2.9.5
pytz==2022.6
sqlparse==0.4.3
